+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Fix more MySQL errors with boolean properties in TYPO3 8.7 (#262)
 
 ## 2.3.2